requirements.txt
# Core runtime
numpy>=1.26,<3
pandas>=2.2,<3
scipy>=1.11,<2
tqdm>=4.66,<5

# ML
# For GPU, install a CUDA-enabled torch build matching your server CUDA.
# Example (CUDA 12.4):
# pip install --index-url https://download.pytorch.org/whl/cu124 torch torchvision torchaudio
torch>=2.2,<3
scikit-learn>=1.4,<2

# Visualization
matplotlib>=3.8,<4
seaborn>=0.13,<1

# API / configs
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
PyYAML>=6.0,<7

# Notebook environment
jupyterlab>=4.2,<5
ipykernel>=6.29,<7
